ApiAutoCore/requirements.txt
allure-pytest==2.8.36
allure-python-commons==2.8.36
configparser==5.0.0
PyMySQL==0.9.3
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.11.0
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
xlrd==1.2.0
xlwt==1.3.0
pyDes==2.0.1
ruamel.yaml==0.17.10